+++ jsonargparse/_loaders_dumpers.py
@@ -329,7 +329,7 @@ def set_omegaconf_loader():
     if omegaconf_support and "omegaconf" not in loaders:
         from ._optionals import get_omegaconf_loader
 
-        set_loader("omegaconf", get_omegaconf_loader())
+        set_loader("omegaconf", get_omegaconf_loader(), get_loader_exceptions("yaml"))
 
 
 set_loader("jsonnet", jsonnet_load, get_loader_exceptions("jsonnet"))